// src/modules/common/types/courses.ts
/**
 * Autogenerated bt Json2Ts (http://json2ts.com/)
 * Change if needed
 */

export interface Term {
  id: number;
  name: string;
  start_at: Date;
  end_at: Date | null;
  created_at: Date;
  workflow_state: string;
  grading_period_group_id?: any;
}

export interface Calendar {
  ics: string;
}

export interface Enrollment {
  type: string;
  role: string;
  role_id: number;
  user_id: number;
  enrollment_state: string;
  limit_privileges_to_course_section: boolean;
}

export interface RESTCanvasCourse {
  id: number;
  root_account_id: number;
  account_id: number;
  name: string;
  enrollment_term_id: number;
  uuid: string;
  start_at: Date;
  grading_standard_id?: any;
  is_public?: any;
  created_at: Date;
  course_code: string;
  default_view: string;
  license?: any;
  grade_passback_setting?: any;
  end_at?: any;
  public_syllabus: boolean;
  public_syllabus_to_auth: boolean;
  storage_quota_mb: number;
  is_public_to_auth_users: boolean;
  homeroom_course: boolean;
  course_color?: any;
  friendly_name?: any;
  term: Term;
  apply_assignment_group_weights: boolean;
  calendar: Calendar;
  time_zone: string;
  blueprint: boolean;
  template: boolean;
  enrollments: Enrollment[];
  hide_final_grades: boolean;
  workflow_state: string;
  restrict_enrollments_to_course_dates: boolean;
  image_download_url?: string;
}